requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
PyYAML==6.0.2
Pillow==11.0.0
prometheus-client==0.20.0
pydantic==2.7.4
